more progress
current issue (i think):
- cant reliably determine the state of the player using
notifications and responses to play and pause requests.
need to use getproperties or something similar to get the
state of each client. this is a problem because state is set
to non busy before the client has notified us of its new state.
this throws off the global pausehandler and pauses/resumes other
clients when they shouldnt be paused or resumed.

other than this, the handlers work separately (i think)