fix client syncing and global pause/resume
this fixes pausing clients that are ahead without
triggering the global pause/play handler used to send
a global pause/play to all clients in the pool.

this leads to the following two features:
- any clients that are ahead of another client will be
paused in a way that matches all the other clients,
effectively syncing the progress of all clients.
- any pause/play "signals" sent from one client will
be broadcasted to all other clients in the pool to
make them perform the same action